middlewares implemented

+++ middlewares/question/deleteQuestion.js
@@ -1,4 +1,5 @@
 const Question = require('../../model/Question');
+const User = require('../../model/User');
 /**
  * Delete question
  */
@@ -11,15 +12,32 @@ module.exports = objectrepository => {
             return next();
         }
 
-        Question.deleteOne({
-            _id: res.locals.question._id
-        }, err => {
-            if (err) {
+        User.findOne({
+            _id: req.session.userid
+        })
+            .populate('_ownQ')
+            .exec()
+            .then(user => {
+
+                for (let i = 0; i < user._ownQ.length; i++) {
+                    if (user._ownQ[i]._id.equals(res.locals.question._id)) {
+                        Question.deleteOne({
+                            _id: res.locals.question._id
+                        }, err => {
+                            if (err) {
+                                console.log(err);
+                            } else {
+                                console.log('OK');
+                            }
+                        });
+                        break;
+                    }
+                }
+            })
+            .catch(err => {
                 console.log(err);
-            } else {
-                console.log('OK');
-            }
-        });
+                res.redirect('/question');
+            });
 
         res.redirect('/question');
     };
