+++ middlewares/user/deleteUser.js
@@ -10,15 +10,20 @@ module.exports = objectrepository => {
             return next();
         }
 
-        User.deleteOne({
-            _id: res.locals.user._id
-        }, err => {
-            if (err) {
-                console.log(err);
-            } else {
-                console.log('OK');
-            }
-        });
+        if (res.locals.user._id.equals(req.session.userid)) {
+            User.deleteOne({
+                _id: res.locals.user._id
+            }, err => {
+                if (err) {
+                    console.log(err);
+                } else {
+                    req.session.destroy(err => {
+                        return next();
+                    });
+                }
+            });
+        }
+
 
         res.redirect('/question');
     };
